Initial Commit
This project is being moved from my private SVN repository to GitHub for public consumption. It is hereby licensed under the Lesser GPL for free and commercial usage as a library to your own project.

In this commit is a README file to provide license information, setup help, and usage details.

The code written so far is by in a large a good amount of work that has taken months to complete and is in no way a final product. Much of the code is undocumented and requires some attention to it's "usability". That being said most of the features present are in working order given an adequate web browser.

A few examples are also added to show case some of the features and use cases for the engine's capabilities.

Enjoy,
Signed-off-by: Ryan Badour <rybadour@gmail.com>