Added: A Clear color for convenience in the list of defined colors in Color.hx
Fixed: Noticed a lack of proper cleanup in the ManagedObject.hx, it was not cleaning up it's register events upon destruction.

Fixed (sorta): The mouse region manager could potentially add handlers for components that weren't initialized and this made them inactive. So it is impossible now to get a handler for an uninitialized component.