Updated: Circle and Rect now inherit from a new base class called Shape, they implement 2 functions used to provide a common interface between their respective functions. Updated: Simplified the collision components slightly to accommodate shape changes.
Fixed: The ui/Button class was not destroying its components and objects.

Added: MouseExclusionRegions can be created in the MouseRegionManager or from MouseHandlers now. They allow the developer to create shapes to block mouse input on mouse handlers below them. To add a sense of depth to the mouse handlers a new property "depth" is added. The value is arbitrary but determines which handlers are blocked if exclusion regions are present.

Signed-off-by: Ryan Badour <rybadour@gmail.com>