Fixed: The ui button classes now work with the new scene code.
Updated: Further work on integrating scenes with the code base. There is now code to support the render depth of scenes that run concurrently. Each scene has a separate canvas for rendering (through the RenderLayerManager.hx) and they overlap exactly inside the target element.

Updated: The global scene was removed due to the fact that scenes are more personal now and a global one isn't necessary when the developer will likely be modifying how it works themself.

Signed-off-by: Ryan Badour <rybadour@gmail.com>