Added: Numerous classes and interfaces to represent components used from rendering the canvas. A basic canvas renderer is something with a render state object that controls the properties all things should have for rendering to a canvas.
Rendering is done by passing a canvas to a renderer so it's no longer dependent on a specific thing to know how to draw.
RendererGroups are designed as a hierarchy of things that render to it's canvas and that canvas onto one provided by what ever tells the group to render.

Signed-off-by: Ryan Badour <rybadour@gmail.com>